(trying to) debugging building under 64bit linux
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5147 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicleControl.h
@@ -79,16 +79,16 @@ public:
     void newUnbuildVehicleBuild();
 
     /// Returns the number of build vehicles
-    size_t getLoadedVehicleNo() const;
+    unsigned int getLoadedVehicleNo() const;
 
     /// Returns the number of removed vehicles
-    size_t getEndedVehicleNo() const;
+    unsigned int getEndedVehicleNo() const;
 
     /// Returns the number of build and emitted, but not yet deleted vehicles
-    size_t getRunningVehicleNo() const;
+    unsigned int getRunningVehicleNo() const;
 
     /// Returns the number of emitted vehicles
-    size_t getEmittedVehicleNo() const;
+    unsigned int getEmittedVehicleNo() const;
 
     /// Return the meaning waiting time of vehicles (corn-dependent value)
     SUMOReal getMeanWaitingTime() const;
